Inline the streambuf bridges into ffi.cpp and slim ffi.h
RustReadStreambuf / RustWriteStreambuf are used only by the ffi.cpp
implementations, so their definitions move out of the header, with
method bodies written inside the class. ffi.h drops <streambuf>,
<cstdint>, <vector>, and <memory>: rust/cxx.h provides cstdint/vector
and the OCCT headers provide memory transitively.
